+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * [Fixed] `autocompleteCountries` prop on `AddressSheet` was silently ignored on iOS.
 * [Fixed] Calling `confirmPayment` without `params` no longer crashes in iOS builds. ([#2331](https://github.com/stripe/stripe-react-native/issues/2331))
 
+**Features**
+* [Added] Added Multibanco support on iOS and Android. ([#2352](https://github.com/stripe/stripe-react-native/issues/2352))
+
 ## 0.63.0 - 2026-04-06
 **Changes**
 - Updated Stripe iOS SDK from 25.9.+ to 25.10.+.